--- Java Obfuscator/.gitignore
@@ -1,9 +1,0 @@
-# User Specific Idea Files
-/.idea/**/workspace.xml
-/.idea/**/tasks.xml
-/.idea/dictionaries
-/.idea/**/dynamic.xml
-/.idea/**/uiDesigner.xml
-
-# Production Files
-**/out
