--- Java Obfuscator/.idea/libraries/jfoenix_8_0_1.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="jfoenix-8.0.1">
-    <CLASSES>
-      <root url="jar://H:/Java Libraries/jfoenix-8.0.1.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
