+++ src/gui/assets/clear-icon.png
