+++ src/gui/assets/copy-icon.png
