+++ src/gui/assets/icon.png
